Delete Invoices screen and ensured everything is working. Next to work on Send screen
